backend/requirements.txt
fastapi==0.110.1
uvicorn==0.25.0

# MongoDB
motor==3.3.1
pymongo==4.5.0

# Validation
pydantic==2.12.4
python-dateutil==2.9.0.post0

# Environment variables
python-dotenv==1.2.1

# Multipart uploads (optional but usually needed)
python-multipart==0.0.20

# Auth (optional, only if you need JWT)
python-jose==3.5.0
passlib==1.7.4
bcrypt==3.2.2

# HTTP requests
requests==2.32.5
httpx==0.28.1

# OpenAI / litellm (your LLM usage)
openai
litellm==1.80.0
tiktoken==0.12.0

# Utils
typing_extensions==4.15.0

langchain-core>=0.1.0,<0.3.0
langchain-openai>=0.0.5
langchain-google-genai>=1.0.0
playwright>=1.40.0
google-generativeai>=0.3.0
aiohttp>=3.9.0